# WebApi.MinimalApi/Models/UserDto.cs
using System;

namespace WebApi.Models
{
    public class UserDto
    { 
        public Guid Id { get; set; }
        public string Login { get; set; }
        public string FullName { get; set; }
        public int GamesPlayed { get; set; }
        public Guid? CurrentGameId { get; set; }
    }
}